CHATBOT-RAG/requirements.txt
Flask
Flask-cors
ollama
faiss-cpu
sentence-transformers
pymupdf
scikit-learn
